fixed the plugin for use in sylius 1.9 - would be better to change the injection of entityRepository to something specific rather than using a decorator to fix that
