+++ src/DependencyInjection/Configuration.php
@@ -14,8 +14,8 @@ final class Configuration implements ConfigurationInterface
      */
     public function getConfigTreeBuilder(): TreeBuilder
     {
-        $treeBuilder = new TreeBuilder();
-        $rootNode = $treeBuilder->root('fos_sylius_import_export');
+        $treeBuilder = new TreeBuilder('');
+        $rootNode = $treeBuilder->getRootNode('fos_sylius_import_export');
 
         $rootNode
             ->children()
